+++ NEWS.md
@@ -1,3 +1,10 @@
+# utsf 1.1.0
+
+* Improvements in estimation of forecast accuracy with rolling origin evaluation.
+* The way in which preprocessings are specified has changed.
+* Method plot.utsf is implemented.
+* Linear models (stats::lm) are supported.
+
 # utsf 1.0.0
 
 * Initial CRAN submission.